requirements.txt
# budAI - RAG API Service
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.1

# AI Providers
openai==1.59.5
google-genai==1.5.0
httpx==0.28.1

# RAG Pipeline
chromadb==0.6.3
langchain-text-splitters==0.3.6

# Document Parsing
pypdf==5.1.0
python-docx==1.1.2
chardet==5.2.0

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0
